Remove double cache + refactor
